Forge check

+++ src/WrappedReact.sol
@@ -6,10 +6,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "./AuthorizedMinter.sol";
 
 contract WrappedReact is ERC20, Ownable, AuthorizedMinter {
-    constructor()
-        ERC20("Wrapped REACT", "wREACT")
-        Ownable(msg.sender)
-    {}
+    constructor() ERC20("Wrapped REACT", "wREACT") Ownable(msg.sender) {}
 
     function authorizeMinter(address minter) public onlyOwner {
         _authorizeMinter(minter);
@@ -18,12 +15,12 @@ contract WrappedReact is ERC20, Ownable, AuthorizedMinter {
     function deAuthorizeMinter(address minter) public onlyOwner {
         _deAuthorizeMinter(minter);
     }
- 
+
     function mint(address to, uint256 amount) public onlyAuthorizedMinter {
         _mint(to, amount);
     }
 
-    function burn(address from, uint amount) public onlyAuthorizedMinter {
+    function burn(address from, uint256 amount) public onlyAuthorizedMinter {
         _burn(from, amount);
     }
 }
